+++ services/cognitiveservices/mgmt/2021-10-01/cognitiveservices/CHANGELOG.md
@@ -1,2 +1,7 @@
-# Change History
+# Unreleased
 
+## Additive Changes
+
+### New Constants
+
+1. DeploymentScaleType.DeploymentScaleTypeStandard